roles/requirements.txt
appnope==0.1.0
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
